Automate audio track extraction and replacement and mod merging and packing
Configurable and extendable python script to automate the process of audio modding in Unreal Engine games with Criware audio banks and a mix of IoStore and legacy packaging.  Sample audio mods and config files for the music and character voice banks in the AEW Fight Forever game are included.  This framework should be extendable to other banks and other games by adding new configuration files in config_files and new uecastoc manifest files in menv_utoc_maifests.

The script automates:
 - Extracting the audio tracks from the game's audio banks (it assumes you already extracted the banks themselves using Fmodel).
 - Merging different audio mods that all touch the same banks
 - Batch conversion of mod tracks in to HCA and batch insertion of modded tracks into the appropriate banks
 - Batch packaging of modded audio banks for both IoStore and legacy packaging modes.

 TODO:
  - No checking whether different audio mods modify the same track of the same bank
  - Add more robust error checking and handling to the script
  - Currently, the AEW generic_roster bank